Reorg the solution structure and put files not part of the lesson in sub solution folder.
