+++ Robot/Game/Direction.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System.Text;
 
-namespace Robot
+namespace Game
 {
     public enum Direction
     {
