+++ Robot/RobotMover.cs
@@ -1,10 +1,6 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
-namespace Robot
+namespace Game
 {
     public class RobotMover
     {
