+++ TutorialsFromBeginning.sln
@@ -16,7 +16,11 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "PersonList", "PersonList\Pe
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "IntList", "IntList\IntList.csproj", "{1FF7BF3B-E051-4976-BB39-960C27881ACE}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Robot", "Robot\Robot.csproj", "{644E9107-CDD5-4D04-944A-83FF7BB4E622}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Robot", "Robot\Robot.csproj", "{644E9107-CDD5-4D04-944A-83FF7BB4E622}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Old", "Old", "{E8971203-1E98-4010-9A28-5755244566ED}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "1-Robot", "1-Robot", "{0668765C-A23D-4E24-A8B0-F92DBC4980BF}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -47,6 +51,13 @@ Global
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(NestedProjects) = preSolution
+		{D052BC73-613B-49C8-BCEA-B2E9F4D8DDD1} = {E8971203-1E98-4010-9A28-5755244566ED}
+		{3383D8F8-3B6A-4154-B908-3B7E186E7AD8} = {E8971203-1E98-4010-9A28-5755244566ED}
+		{0C3610FC-88D5-47FE-8206-A32CD82878AB} = {E8971203-1E98-4010-9A28-5755244566ED}
+		{1FF7BF3B-E051-4976-BB39-960C27881ACE} = {E8971203-1E98-4010-9A28-5755244566ED}
+		{644E9107-CDD5-4D04-944A-83FF7BB4E622} = {0668765C-A23D-4E24-A8B0-F92DBC4980BF}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {BAA62E8A-32E7-49AF-9CAA-10653C51841D}
